Update examples to use direct functions
Direct functions are now the standard across the examples. The former *_direct_functions.py variants become the canonical food_ordering.py and restaurant_reservation.py, the remaining schema-based examples are converted, and one FlowsFunctionSchema example is kept as food_ordering_advanced_functionschema.py to show the explicit-schema approach (property control like enum and min/max constraints).
