+++ examples/podcast_interview.py
@@ -48,12 +48,7 @@ from pipecat.transports.daily.transport import DailyParams
 from pipecat.workers.runner import WorkerRunner
 from utils import create_llm
 
-from pipecat_flows import (
-    FlowArgs,
-    FlowManager,
-    FlowsFunctionSchema,
-    NodeConfig,
-)
+from pipecat_flows import FlowManager, NodeConfig
 
 load_dotenv(override=True)
 
@@ -71,28 +66,49 @@ class StartInterviewResult(TypedDict):
     topic: str
 
 
+# Functions for Introduction Node
+async def proceed_to_topic(
+    flow_manager: FlowManager, guest_summary: str
+) -> tuple[ProceedToTopicResult | None, NodeConfig]:
+    """Use after the guest has introduced themselves.
+
+    Args:
+        guest_summary (str): A quick summary of who the guest is (name, role, area of expertise, etc.).
+    """
+    return ProceedToTopicResult(guest_summary=guest_summary), create_topic_node()
+
+
+# Functions for Topic Node
+async def start_interview(
+    flow_manager: FlowManager, topic: str
+) -> tuple[StartInterviewResult | None, NodeConfig]:
+    """Use this when the guest has shared a clear topic they want to explore.
+
+    Args:
+        topic (str): The topic the guest wants to discuss.
+    """
+    return StartInterviewResult(topic=topic), create_interview_node()
+
+
+# Functions for Interview Node
+async def next_question(flow_manager: FlowManager) -> tuple[None, NodeConfig]:
+    """Use this after you've thoroughly explored the current aspect with multiple questions and follow-ups."""
+    return None, create_interview_node()
+
+
+async def wrap_up(flow_manager: FlowManager) -> tuple[None, NodeConfig]:
+    """Use this when you've gathered substantial insights and are ready to wrap up."""
+    return None, create_conclusion_node()
+
+
+# Functions for Conclusion Node
+async def end_interview(flow_manager: FlowManager) -> tuple[None, NodeConfig]:
+    """Use this after the guest has shared their final thoughts."""
+    return None, create_final_node()
+
+
 def create_introduction_node() -> NodeConfig:
     """Create the Introduction node."""
-
-    async def handle_proceed_to_topic(
-        args: FlowArgs, flow_manager: FlowManager
-    ) -> tuple[ProceedToTopicResult | None, NodeConfig]:
-        """Handler for proceed_to_topic function"""
-        guest_summary: str = args.get("guest_summary", "")
-        return ProceedToTopicResult(guest_summary=guest_summary), create_topic_node()
-
-    proceed_to_topic_func = FlowsFunctionSchema(
-        name="proceed_to_topic",
-        handler=handle_proceed_to_topic,
-        description="Use after the guest has introduced themselves.",
-        properties={
-            "guest_summary": {
-                "type": "string",
-                "description": "A quick summary of who the guest is (name, role, area of expertise, etc.)",
-            }
-        },
-        required=["guest_summary"],
-    )
     return NodeConfig(
         name="introduction",
         role_message="You are a warm, engaging podcast host with a natural conversational style. You're genuinely curious about your guests and skilled at making them feel comfortable while drawing out interesting insights. Your questions flow naturally, and you listen actively, building on what your guest shares.",
@@ -102,29 +118,12 @@ def create_introduction_node() -> NodeConfig:
                 "content": "Welcome the guest warmly and enthusiastically. Focus this exchange on getting to know who they are. Invite them to briefly introduce themselves—name, role, current focus, or anything fun they'd like to share. Ask one follow-up question if it helps clarify or highlight something interesting about them. Once you feel you have a clear introduction, use the proceed_to_topic function to move into topic selection.",
             }
         ],
-        functions=[proceed_to_topic_func],
+        functions=[proceed_to_topic],
     )
 
 
 def create_topic_node() -> NodeConfig:
     """Create the Topic Selection node."""
-
-    async def handle_start_interview(
-        args: FlowArgs, flow_manager: FlowManager
-    ) -> tuple[StartInterviewResult | None, NodeConfig]:
-        """Handler for start_interview function"""
-        topic: str = args.get("topic", "")
-        return StartInterviewResult(topic=topic), create_interview_node()
-
-    start_interview_func = FlowsFunctionSchema(
-        name="start_interview",
-        handler=handle_start_interview,
-        description="Use this when the guest has shared a clear topic they want to explore.",
-        properties={
-            "topic": {"type": "string", "description": "The topic the guest wants to discuss"}
-        },
-        required=["topic"],
-    )
     return NodeConfig(
         name="topic",
         task_messages=[
@@ -133,37 +132,12 @@ def create_topic_node() -> NodeConfig:
                 "content": "Now that you know who the guest is, help them choose the topic they'd like to explore. Refer back to their introduction to personalize the transition. Ask what topic, story, or challenge they're excited to discuss today. Show genuine interest and, if needed, ask a clarifying question to make sure you understand the angle they want to take. Once the topic feels clear and specific enough to dive into, use the start_interview function.",
             }
         ],
-        functions=[start_interview_func],
+        functions=[start_interview],
     )
 
 
 def create_interview_node() -> NodeConfig:
     """Create the Interview node."""
-
-    async def handle_next_question(
-        args: FlowArgs, flow_manager: FlowManager
-    ) -> tuple[None, NodeConfig]:
-        """Handler for next_question function"""
-        return None, create_interview_node()
-
-    async def handle_wrap_up(args: FlowArgs, flow_manager: FlowManager) -> tuple[None, NodeConfig]:
-        """Handler for wrap_up function"""
-        return None, create_conclusion_node()
-
-    next_question_func = FlowsFunctionSchema(
-        name="next_question",
-        handler=handle_next_question,
-        description="Use this after you've thoroughly explored the current aspect with multiple questions and follow-ups.",
-        properties={},
-        required=[],
-    )
-    wrap_up_func = FlowsFunctionSchema(
-        name="wrap_up",
-        handler=handle_wrap_up,
-        description="Use this when you've gathered substantial insights and are ready to wrap up.",
-        properties={},
-        required=[],
-    )
     return NodeConfig(
         name="interview",
         task_messages=[
@@ -172,26 +146,12 @@ def create_interview_node() -> NodeConfig:
                 "content": "You're now in the heart of the interview. Start by introducing the topic with enthusiasm, then dive deep into one key aspect at a time. Ask open-ended, thoughtful questions that invite storytelling and personal insights. Listen actively to responses and ask natural follow-up questions that build on what your guest shares—dig deeper into interesting points, ask for examples, or explore the 'why' behind their answers. Keep the conversation flowing naturally, like a genuine dialogue between friends. Once you've thoroughly explored an aspect (typically after 3-5 exchanges), use the next_question function to smoothly transition to the next key aspect. After covering 3 key aspects of the topic, use the wrap_up function to conclude the interview.",
             }
         ],
-        functions=[next_question_func, wrap_up_func],
+        functions=[next_question, wrap_up],
     )
 
 
 def create_conclusion_node() -> NodeConfig:
     """Create the Conclusion node."""
-
-    async def handle_end_interview(
-        args: FlowArgs, flow_manager: FlowManager
-    ) -> tuple[None, NodeConfig]:
-        """Handler for end_interview function"""
-        return None, create_final_node()
-
-    end_interview_func = FlowsFunctionSchema(
-        name="end_interview",
-        handler=handle_end_interview,
-        description="Use this after the guest has shared their final thoughts.",
-        properties={},
-        required=[],
-    )
     return NodeConfig(
         name="conclusion",
         task_messages=[
@@ -200,7 +160,7 @@ def create_conclusion_node() -> NodeConfig:
                 "content": "Express genuine appreciation for the conversation and the insights your guest shared. Summarize 2-3 key takeaways or memorable points from your discussion in a warm, conversational way—this helps reinforce the value of the conversation. Then, ask your guest if they have any final thoughts, a last word, or anything else they'd like to add. Wait for their response, then use the end_interview function to wrap up.",
             }
         ],
-        functions=[end_interview_func],
+        functions=[end_interview],
     )
